refactor: rename file `nginx.ts` => `ingress-controller.ts`
Let the name be more specific.
