+++ core/src/plugins/kubernetes/nginx/ingress-controller.ts
